% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_dist_seeker_gpg.R
\name{min.dist.seeker.gpg}
\alias{min.dist.seeker.gpg}
\title{Find minimal distance}
\usage{
\method{min}{dist.seeker.gpg}(data.1, data.2, a.min = 0.1, a.max = 1.9,
  a.skok = 0.1, theta.min = -0.9, theta.max = 0.9, theta.skok = 0.1,
  curves = 0, wydruk = T)
}
\arguments{
\item{data.1}{First specimen}

\item{data.2}{Second specimen}

\item{a.min}{a.min}

\item{a.max}{a.max}

\item{a.skok}{a.skok}

\item{theta.min}{theta.min}

\item{theta.max}{theta.max}

\item{theta.skok}{theta.skok}

\item{curves}{Curve matrix for semilandmarks sliding}

\item{wydruk}{Print distance}

\item{iteracje}{How precize should the search be}

\item{istotne.cyfry}{Digit number in comparing distance}
}
\description{
Function to find minimal Procrustes distance between two specimens
}

