% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimal_distance_matrix_iter_half.R
\name{minimal.distance.matrix.iter.half}
\alias{minimal.distance.matrix.iter.half}
\title{Create minimal distance matrix}
\usage{
minimal.distance.matrix.iter.half(data, Csize = FALSE, curve = 0)
}
\arguments{
\item{data}{Landmark dataset}

\item{Csize}{Optional column with centroid size. Defaults to FALSE}

\item{curve}{Optional matrix for sliding semilandmarks}
}
\description{
Function for generating minimal Procrustes distance matrix from tectonicaly deformed landmark dataset.
}

