% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_minimal_distance.R
\name{partial.minimal.distance}
\alias{partial.minimal.distance}
\title{Calculate procrustes distance between parts}
\usage{
partial.minimal.distance(data, list.of.elements, a, theta, distance = "full",
  curves = NULL)
}
\arguments{
\item{data}{landmark array}

\item{list.of.elements}{list of elements}

\item{a}{a matrix}

\item{theta}{theta matrix}

\item{distance}{distance type ('full' or 'procrustes')}

\item{curves}{Curve matrix for semilandmarks sliding}
}
\description{
Calculate procrustes distance between parts of objects
}

