% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultant_distance_matrix.R
\name{resultant.distance.matrix}
\alias{resultant.distance.matrix}
\title{Calcualte resultant distance matrix}
\usage{
resultant.distance.matrix(..., method)
}
\arguments{
\item{...}{partial distance matrices}

\item{method}{method of calculation (sum, mean, vector)}
}
\description{
Calculate resultant distance matrix from partial distance matrices.
}

