% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultant_distance.R
\name{resultant.distance}
\alias{resultant.distance}
\title{Calculate resultant distance}
\usage{
resultant.distance(data, method)
}
\arguments{
\item{data}{vector of distances}

\item{method}{method of calculation (sum, mean or vector)}
}
\description{
Calculate resultant distance from several partial distances as sum, mean or resultant vector
}

