\name{RetroGeoMorph-package}
\alias{RetroGeoMorph-package}
\alias{RetroGeoMorph}
\docType{package}
\title{
\packageTitle{RetroGeoMorph}
}
\description{
\packageDescription{RetroGeoMorph}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RetroGeoMorph}
\packageIndices{RetroGeoMorph}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{RetroGeoMorph}

Maintainer: \packageMaintainer{RetroGeoMorph}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
